requirements.txt
flask-restplus==0.11.0
Flask-SQLAlchemy==2.1
werkzeug
flask_jwt_extended
